add gates and policy

+++ public/app/Http/Controllers/Admin/GroupsController.php
@@ -4,8 +4,11 @@ namespace App\Http\Controllers\Admin;
 
 use App\Models\Group;
 use App\Models\Flow;
+use App\User;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Auth;
+use Gate;
 
 class GroupsController extends Controller
 {
@@ -17,7 +20,13 @@ class GroupsController extends Controller
     public function index()
     {
 //        start_measure('start111');
-        $groups = Group::orderBy('created_at', 'desc')->paginate(10);
+
+       // $groups = Group::orderBy('created_at', 'desc')->where('created_by', Auth::id())->paginate(10);
+
+
+        $groups = User::find(Auth::id())->groups()->paginate(10);
+
+
         //$groups->load(['responsibilities']);
         return view('groups.index', [
             'groups' => $groups,
@@ -43,7 +52,13 @@ class GroupsController extends Controller
      */
     public function store(Request $request)
     {
+        $userID = Auth::id();
+        $request['created_by'] = $userID;
+        //return $request;
         $group = Group::create($request->all());
+
+        $group->users()->attach($userID);
+
         return redirect()->route('admin.groups.index');
     }
 
@@ -57,9 +72,10 @@ class GroupsController extends Controller
     {
         return view('groups.show', [
             'group' => $group,
-            'responsibilities'=> $group->load(['responsibilities']),
-            'reasons'=>$group->load(['reasons']),
-            'flows'=> Flow::orderBy('created_at', 'desc')->where('group_id', $group->id)->limit(10)->get()
+            'responsibilities' => $group->load(['responsibilities']),
+            'reasons' => $group->load(['reasons']),
+            'flows' => Flow::orderBy('created_at', 'desc')->where('group_id', $group->id)->limit(10)->get(),
+            'users' => Group::find($group->id)->users()->get(),
         ]);
     }
 
@@ -71,6 +87,10 @@ class GroupsController extends Controller
      */
     public function edit(Group $group)
     {
+        if (Gate::denies('is-owner', $group)) {
+            return 'нет прав';
+        }
+
         return view('groups.edit', [
             'group' => $group
         ]);
@@ -85,6 +105,10 @@ class GroupsController extends Controller
      */
     public function update(Request $request, Group $group)
     {
+        if (Gate::denies('is-owner', $group)) {
+            return 'нет прав';
+        }
+
         $group->update($request->all());
         return redirect()->route('admin.groups.index');
     }
@@ -97,7 +121,40 @@ class GroupsController extends Controller
      */
     public function destroy(Group $group)
     {
+        if (Gate::denies('is-owner', $group)) {
+            return 'нет прав';
+        }
+
         $group->delete();
         return redirect()->route('admin.groups.index');
+    }
+
+    /**
+     * List users for invite to group
+     *
+     * @param Group $group
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     */
+    public function invite(Group $group)
+    {
+
+        $users = User::get();
+
+        return view('groups.invite', [
+            'group' => $group,
+            'users' => $users]);
+    }
+
+    /**
+     *Add User to Group
+     *
+     */
+    public function addUser(Request $request, Group $group)
+    {
+
+        if ($request->input('user_id')):
+            $group->users()->attach($request->input('user_id'));
+        endif;
+        return redirect()->back()->with('message', 'пользователь добавлен');
     }
 }
